% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_transcriptID.R
\name{convert_transcriptID}
\alias{convert_transcriptID}
\title{Convert transcript IDs between different databases}
\usage{
convert_transcriptID(dat, db)
}
\arguments{
\item{dat}{a dataframe including Transcript_version, Gene, exon, Nucleotide, AA_changes}

\item{db}{`EnsDb` object.}
}
\value{
a new dataset with converting information
}
\description{
Preprocessing, mapping/converting, searching SNP for data
}
\examples{
library(EnsDb.Hsapiens.v86)
db=EnsDb.Hsapiens.v86
dat<-read.csv(system.file("extdata",
                          "variant_list_test.csv",
                          package = "CvtransctiptID"),
              stringsAsFactors = FALSE, encoding = "UTF-8", row.names = NULL, sep = ",")
new_dat<-convert_transcriptID(dat, db)

}
